refactor : MainActivity에서 MainViewModel 주입 방식을 ViewModelFactoryInjector를 사용하도록 변경
